Sprint 23

+++ app/src/main/java/com/example/playlistmaker/media/data/db/dao/FavoriteTrackDao.kt
@@ -14,11 +14,11 @@ interface FavoriteTrackDao {
     suspend fun insertTrack(track: FavoriteTrackEntity)
 
     @Query("DELETE FROM favorite_tracks WHERE trackId = :trackId")
-    suspend fun deleteTrack(trackId: Int?)
+    suspend fun deleteTrack(trackId: Long?)
 
     @Query("SELECT * FROM favorite_tracks ORDER BY timestamp DESC")
     fun getTracks(): Flow<List<FavoriteTrackEntity>>
 
     @Query("SELECT trackId FROM favorite_tracks")
-    suspend fun getTracksId(): List<Int>
+    suspend fun getTracksId(): List<Long>
 }
