+++ app/src/main/java/com/example/playlistmaker/media/domain/models/Track.kt
@@ -10,7 +10,7 @@ import java.util.Locale
 
 @Parcelize
 data class Track(
-    val trackId: Int?,
+    val trackId: Long?,
     val trackName: String?,
     val artistName: String?,
     @SerializedName("trackTimeMillis") val trackTime: Long?,
